# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export interface Medicine {
  id: number;
  name: string;
  genericName: string;
  /** @nullable */
  barcode?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierName?: string | null;
  quantity: number;
  minQuantity: number;
  purchasePrice: number;
  sellingPrice: number;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  requiresPrescription?: boolean;
  createdAt: string;
}

export interface MedicineInput {
  /** @minLength 1 */
  name: string;
  genericName: string;
  barcode?: string;
  categoryId?: number;
  supplierId?: number;
  quantity: number;
  minQuantity: number;
  purchasePrice: number;
  sellingPrice: number;
  expiryDate?: string;
  location?: string;
  description?: string;
  requiresPrescription?: boolean;
}

export interface MedicineUpdate {
  name?: string;
  genericName?: string;
  barcode?: string;
  categoryId?: number;
  supplierId?: number;
  quantity?: number;
  minQuantity?: number;
  purchasePrice?: number;
  sellingPrice?: number;
  expiryDate?: string;
  location?: string;
  description?: string;
  requiresPrescription?: boolean;
}

export interface Supplier {
  id: number;
  name: string;
  /** @nullable */
  contactPerson?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt: string;
}

export interface SupplierInput {
  /** @minLength 1 */
  name: string;
  contactPerson?: string;
  phone?: string;
  email?: string;
  address?: string;
}

export interface SupplierUpdate {
  name?: string;
  contactPerson?: string;
  phone?: string;
  email?: string;
  address?: string;
}

export interface PurchaseItem {
  id?: number;
  medicineId: number;
  medicineName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  /** @nullable */
  expiryDate?: string | null;
}

export interface PurchaseItemInput {
  medicineId: number;
  quantity: number;
  unitPrice: number;
  expiryDate?: string;
}

export type PurchaseStatus = typeof PurchaseStatus[keyof typeof PurchaseStatus];


export const PurchaseStatus = {
  pending: 'pending',
  received: 'received',
  cancelled: 'cancelled',
} as const;

export interface Purchase {
  id: number;
  supplierId: number;
  supplierName: string;
  /** @nullable */
  invoiceNumber?: string | null;
  totalAmount: number;
  status: PurchaseStatus;
  /** @nullable */
  notes?: string | null;
  items: PurchaseItem[];
  createdAt: string;
}

export interface PurchaseInput {
  supplierId: number;
  invoiceNumber?: string;
  notes?: string;
  items: PurchaseItemInput[];
}

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  notes?: string | null;
  totalPurchases?: number;
  createdAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  phone?: string;
  email?: string;
  address?: string;
  dateOfBirth?: string;
  notes?: string;
}

export interface CustomerUpdate {
  name?: string;
  phone?: string;
  email?: string;
  address?: string;
  dateOfBirth?: string;
  notes?: string;
}

export interface SaleItem {
  id?: number;
  medicineId: number;
  medicineName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export interface SaleItemInput {
  medicineId: number;
  quantity: number;
  unitPrice: number;
}

export type SalePaymentMethod = typeof SalePaymentMethod[keyof typeof SalePaymentMethod];


export const SalePaymentMethod = {
  cash: 'cash',
  card: 'card',
  insurance: 'insurance',
} as const;

export type SaleStatus = typeof SaleStatus[keyof typeof SaleStatus];


export const SaleStatus = {
  completed: 'completed',
  refunded: 'refunded',
  pending: 'pending',
} as const;

export interface Sale {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  prescriptionId?: number | null;
  totalAmount: number;
  discount?: number;
  paymentMethod: SalePaymentMethod;
  status: SaleStatus;
  /** @nullable */
  notes?: string | null;
  items: SaleItem[];
  createdAt: string;
}

export type SaleInputPaymentMethod = typeof SaleInputPaymentMethod[keyof typeof SaleInputPaymentMethod];


export const SaleInputPaymentMethod = {
  cash: 'cash',
  card: 'card',
  insurance: 'insurance',
} as const;

export interface SaleInput {
  customerId?: number;
  prescriptionId?: number;
  items: SaleItemInput[];
  discount?: number;
  paymentMethod: SaleInputPaymentMethod;
  notes?: string;
}

export type PrescriptionStatus = typeof PrescriptionStatus[keyof typeof PrescriptionStatus];


export const PrescriptionStatus = {
  pending: 'pending',
  dispensed: 'dispensed',
  cancelled: 'cancelled',
} as const;

export interface Prescription {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  doctorName: string;
  /** @nullable */
  doctorSpecialty?: string | null;
  status?: PrescriptionStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PrescriptionInput {
  customerId?: number;
  /** @minLength 1 */
  doctorName: string;
  doctorSpecialty?: string;
  notes?: string;
}

export type PrescriptionUpdateStatus = typeof PrescriptionUpdateStatus[keyof typeof PrescriptionUpdateStatus];


export const PrescriptionUpdateStatus = {
  pending: 'pending',
  dispensed: 'dispensed',
  cancelled: 'cancelled',
} as const;

export interface PrescriptionUpdate {
  doctorName?: string;
  doctorSpecialty?: string;
  status?: PrescriptionUpdateStatus;
  notes?: string;
}

export interface DashboardSummary {
  todaySales: number;
  todayRevenue: number;
  totalMedicines: number;
  lowStockCount: number;
  expiringSoonCount: number;
  totalCustomers: number;
  monthRevenue: number;
  totalSalesToday: number;
}

export interface SalesChartPoint {
  date: string;
  revenue: number;
  salesCount: number;
}

export interface TopMedicine {
  medicineId: number;
  medicineName: string;
  totalQuantity: number;
  totalRevenue: number;
}

export type ListMedicinesParams = {
search?: string;
categoryId?: number;
lowStock?: boolean;
};

export type ListCustomersParams = {
search?: string;
};

export type ListSalesParams = {
startDate?: string;
endDate?: string;
};