# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Purchases ────────────────────────────────────────────────
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List all purchase orders
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"

  # ─── Customers ────────────────────────────────────────────────
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale (POS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # ─── Prescriptions ────────────────────────────────────────────
...
components:
  schemas:
...
    Supplier:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string

    # ── Purchase ──────────────────────────────────────────────────
    PurchaseItem:
      type: object
      required: [medicineId, medicineName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        expiryDate:
          type: ["string", "null"]
    PurchaseItemInput:
      type: object
      required: [medicineId, quantity, unitPrice]
      properties:
        medicineId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        expiryDate:
          type: string
    Purchase:
      type: object
      required: [id, supplierId, supplierName, totalAmount, status, items, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        invoiceNumber:
          type: ["string", "null"]
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, received, cancelled]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        createdAt:
          type: string
    PurchaseInput:
      type: object
      required: [supplierId, items]
      properties:
        supplierId:
          type: integer
        invoiceNumber:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItemInput"

    # ── Customer ──────────────────────────────────────────────────
...
    SaleItem:
      type: object
      required: [medicineId, medicineName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
    SaleItemInput:
      type: object
      required: [medicineId, quantity, unitPrice]
      properties:
        medicineId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
    Sale:
      type: object
      required: [id, totalAmount, paymentMethod, status, items, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        prescriptionId:
          type: ["integer", "null"]
        totalAmount:
          type: number
        discount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, insurance]
        status:
          type: string
          enum: [completed, refunded, pending]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
    SaleInput:
      type: object
      required: [items, paymentMethod]
      properties:
        customerId:
          type: integer
        prescriptionId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
        discount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, insurance]
        notes:
          type: string

    # ── Prescription ──────────────────────────────────────────────
...