# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierUpdate {
  name?: string;
  contactPerson?: string;
  phone?: string;
  email?: string;
  address?: string;
}